Split utilities into courseUtilities
