+++ frontend/src/components/WorksheetToggleButton.js
@@ -5,10 +5,10 @@ import { Button, Tooltip, OverlayTrigger } from 'react-bootstrap';
 import axios from 'axios';
 import { useUser } from '../user';
 import { toast } from 'react-toastify';
-import { isInWorksheet } from '../utilities';
 import posthog from 'posthog-js';
 import styled from 'styled-components';
 import { getSSObject, setSSObject } from '../browserStorage';
+import { isInWorksheet } from '../courseUtilities';
 
 /**
  * Render worksheet list in default worksheet view
